+++ packages/react-wrap-with/__tests__/__types__/extractProps.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import wrapWith from '../../src/wrapWith';
+
+import type { PropsWithChildren } from 'react';
+
+const Header = ({ children, className }: PropsWithChildren<{ className: string }>) => (
+  <h1 className={className}>{children}</h1>
+);
+
+const Component = wrapWith(Header, {}, ['className'])(() => <div />);
+
+<Component className="123" />;