+++ packages/react-wrap-with/__tests__/__types__/failing/extractProps.emptyProp.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+import wrapWith from '../../../src/wrapWith';
+
+import type { ReactNode } from 'react';
+
+const Header = ({ children }: { children?: ReactNode | undefined }) => <h1>{children}</h1>;
+
+// "xyz" is not a prop of <Header> and cannot be extracted.
+wrapWith(Header, {}, ['xyz']);