+++ packages/react-wrap-with/__tests__/wrapWith.typing.spec.tsx
@@ -30,13 +30,17 @@ function compile(...filenames) {
 }
 
 test('should succeed when mergeProps argument is not passed on component without props', () =>
-  compile('./__types__/noProps.tsx'));
+  compile('./__types__/emptyProps.tsx'));
 
 test('should succeed when undefined is passed to component without props', () =>
-  compile('./__types__/noProps.undefined.tsx'));
+  compile('./__types__/emptyProps.undefined.tsx'));
 
 test('should succeed when empty object is passed to component without props', () =>
-  compile('./__types__/noProps.empty.tsx'));
+  compile('./__types__/emptyProps.empty.tsx'));
+
+test('should succeed when rendering without an extracted optional props', () => compile('./__types__/extractProps.optionalProp.tsx'));
+
+test('should succeed when extracting props', () => compile('./__types__/extractProps.tsx'));
 
 test.failing('should fail when mergeProps is not passed on component requires props', () =>
   compile('./__types__/failing/requireProps.tsx')
@@ -48,4 +52,22 @@ test.failing('should fail when props is passed to component which do not need pr
   compile('./__types__/failing/extraneousProps.tsx')
 );
 
-test.failing('should fail when wrapper component does not allow children', () => compile('./__types__/failing/noChildren.tsx'));
+test.failing('should fail when wrapper component does not allow children', () =>
+  compile('./__types__/failing/noChildren.tsx')
+);
+
+test.failing('should fail when initializing with an extracted prop', () =>
+  compile('./__types__/failing/extractProps.cannotSetInitialProps.tsx')
+);
+
+test.failing('should fail when extracting a non-existing prop from a wrapper component without any props', () =>
+  compile('./__types__/failing/extractProps.emptyProp.tsx')
+);
+
+test.failing('should fail when not rendering with extracted required prop', () =>
+  compile('./__types__/failing/extractProps.missingExtractedRequiredProp.tsx')
+);
+
+test.failing('should fail when extracting a non-existing prop', () =>
+  compile('./__types__/failing/extractProps.noExtractPropKey.tsx')
+);